Fix ballpark_azimuth_angle unmarshal crash on fractional values
ballpark.azimuthAngle was typed int32 based on a single sample date
where it happened to be whole; on 2026-07-31 the API returns it as a
float (43.61 for Angel Stadium), which fails JSON unmarshal into an
int32 field. Retyped to float32 (jsonresponse.Ballpark.AzimuthAngle,
model.Weather.BallparkAzimuthAngle) -- every other ballpark numeric
field stayed a clean int across a 25-game sample across both dates,
so left as-is; azimuthAngle is a computed geographic bearing rather
than a rounded distance/count, which explains why only it varies.

Added TestWeatherScraper_FractionalBallparkAzimuthAngle as a
regression guard against this specific date/game.

+++ dataprovider/propfinder/mlb/jsonresponse/weather.go
@@ -15,7 +15,7 @@ type Ballpark struct {
 	Name              string  `json:"name"`
 	Latitude          float32 `json:"latitude"`
 	Longitude         float32 `json:"longitude"`
-	AzimuthAngle      int32   `json:"azimuthAngle"`
+	AzimuthAngle      float32 `json:"azimuthAngle"`
 	Elevation         int32   `json:"elevation"`
 	Capacity          int32   `json:"capacity"`
 	TurfType          string  `json:"turfType"`
